LPS-144362 search-experiences: Implementation classes go in implementation modules. Only moved, no changes.
https://issues.liferay.com/browse/LPS-144362
